requirements.txt
requests==2.31.0
moviepy==1.0.3
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-api-python-client==2.98.0